Removing default implementations from `HttpClientBuilder`
Motivation:

In order to unify the way methods in `HttpClientBuilder`
and `HttpServerBuilder` are implemented, the default methods from
`HttpClientBuilder` have been removed and implemented in
`SingleAddressHttpClientBuilder`.

Modifications:

- Removed default methods from `HttpClientBuilder`,
- Added implementations to `SingleAddressHttpClientBuilder`,
- Moved utilities to servicetalk-http-netty package,
- Moved tests around,
- Copied methods from `FilterFactoryUtils` to `ConditionalFilterFactory`
  as they are inaccessible from the netty module's testFixtures
  directory.

Result:

The API is consistent with the `HttpServerBuilder`.
